[FIX] my_arraylen switched to size_t return type

+++ lib/my/include/my_arrays.h
@@ -7,5 +7,5 @@
 
 #pragma once
 
-int my_arraylen(char * const *array);
+size_t my_arraylen(char * const *array);
 void my_free_arrays(int size, ...);
